refactoring – move 'create_logprobs' for completion out of serving_engine.py
